--- src/components/RegisterForm/RegisterForm.styles.js
@@ -1,18 +1,0 @@
-import { Formik, Form } from 'formik';
-import Input from '@mui/material/Input';
-import FormControl from '@mui/material/FormControl';
-import styled from '@emotion/styled';
-
-export const StyledForm = styled(Form)`
-    display: flex;
-    flex-direction: column;
-`;
-
-export const StyledFormControl = styled(FormControl)`
-    display: flex;
-    flex-direction: column;
-`;
-
-export const StyledInput = styled(Input)`
-    width: 410px;
-`;
